refactor: .mjs → .ts
