+++ code/sys/sys.data.text/src/Diff/index.ts
@@ -1,1 +1,1 @@
-export * from './Diff.mjs';
+export * from './Diff';
