+++ code/sys/sys.data.text/src/Fuzzy/index.ts
@@ -1,1 +1,1 @@
-export * from './Fuzzy.mjs';
+export * from './Fuzzy';
