+++ code/sys/sys.data.text/src/Markdown.Processor/Processor.ts
@@ -1,5 +1,5 @@
-import { t, trimToPosition } from './common.mjs';
-import { Pipeline } from './util.Pipeline.mjs';
+import { t, trimToPosition } from './common';
+import { Pipeline } from './util.Pipeline';
 
 /**
  * Markdown transformer.
