+++ code/sys/sys.data.text/src/Markdown.Processor/index.ts
@@ -1,1 +1,1 @@
-export { MarkdownProcessor } from './Processor.mjs';
+export { MarkdownProcessor } from './Processor';
