+++ code/sys/sys.data.text/src/Markdown.Processor/types/common.ts
