+++ code/sys/sys.data.text/src/Markdown.Processor/types/types.code.ts
@@ -1,4 +1,4 @@
-import { type t } from './common.mjs';
+import type { t } from './common';
 
 /**
  * Represents a code-block within Markdown.
