+++ code/sys/sys.data.text/src/Markdown.Processor/types/types.mutate.ts
@@ -1,4 +1,4 @@
-import { type t } from './common.mjs';
+import type { t } from './common';
 
 type O = Record<string, unknown>;
