+++ code/sys/sys.data.text/src/Markdown.Processor/util.plugin.CodeBlock.ts
@@ -1,5 +1,5 @@
 import { SKIP, visit } from 'unist-util-visit';
-import { slug, t } from './common.mjs';
+import { slug, t } from './common';
 
 /**
  * Tools for working with tripple-tick (```) code blocks within markdown.
