--- code/sys/sys.data.text/src/Markdown/Markdown.mts
@@ -1,10 +1,0 @@
-import { type t } from '../common';
-import { MarkdownIs as Is } from './Markdown.Is.mjs';
-import { MarkdownProcessor as processor } from '../Markdown.Processor';
-import { MarkdownFind as Find } from './Markdown.Find.mjs';
-
-export const Markdown: t.Markdown = {
-  processor,
-  Is,
-  Find,
-};
