+++ code/sys/sys.data.text/src/Markdown/index.ts
@@ -1,1 +1,1 @@
-export * from './Markdown.mjs';
+export * from './Markdown';
