+++ code/sys/sys.data.text/src/Text.Processor/index.ts
@@ -1,1 +1,1 @@
-export { TextProcessor } from './TextProcessor.mjs';
+export { TextProcessor } from './TextProcessor';
