+++ code/sys/sys.data.text/src/Text/Text.Is.ts
