+++ code/sys/sys.data.text/src/Text/Text.ts
@@ -1,11 +1,11 @@
 import { t, Yaml } from '../common';
-import { TextProcessor as Processor } from '../Text.Processor';
-import { Is } from './Text.Is.mjs';
-import { Markdown } from '../Markdown';
 import { Fuzzy } from '../Fuzzy';
+import { Markdown } from '../Markdown';
+import { TextProcessor as Processor } from '../Text.Processor';
+import { Is } from './Text.Is';
 
 /**
- * Module Root Index
+ * API
  */
 export const Text: t.Text = {
   Is,
