+++ code/sys/sys.data.text/src/common/index.ts
@@ -1,5 +1,5 @@
 export type * as t from './t';
 
 export * from '../index.pkg.mjs';
-export * from './libs.mjs';
-export * from './util.mjs';
+export * from './libs';
+export * from './util';
