+++ code/sys/sys.data.text/src/common/util.ts
