+++ code/sys/sys.data.text/src/global.d.ts
